Adds device settings through protocol buffers.
For now, moves MQTT server address, timezone and node name to it (removed from menuconfig).

Also adds a settings loader to load the settings from am embedded file.
While the settings are embedded for now, the plan is to eventually send
these from a Winston Core server so we can modify them dynamically.